+++ src/imio/helpers/content.py
@@ -216,11 +216,14 @@ def create(conf, cids={}, globl=False, pos=False, clean_globl=False):
     return cids_l
 
 
-def richtextval(text, outputMimeType='text/html'):
+def richtextval(text, mimeType=u"text/html", outputMimeType=u"text/x-html-safe"):
     """
         Return a RichTextValue to be stored in IRichText field
     """
-    return RichTextValue(raw=safe_unicode(text), mimeType='text/html', outputMimeType=outputMimeType, encoding='utf-8')
+    return RichTextValue(raw=safe_unicode(text),
+                         mimeType=mimeType,
+                         outputMimeType=outputMimeType,
+                         encoding='utf-8')
 
 
 @api.validation.at_least_one_of('obj', 'type_name')